generalize the table update, create a new 'IcebergAddSnapshot' for every insert/update/delete
