+++ src/include/iceberg_multi_file_list.hpp
@@ -170,7 +170,7 @@ public:
 	//! The data files of the manifest file that we last scanned
 	idx_t data_file_idx = 0;
 	vector<IcebergManifestEntry> current_data_files;
-	bool scanned_transaction_data = false;
+	idx_t transaction_data_idx = 0;
 
 	//! For each file that has a delete file, the state for processing that/those delete file(s)
 	mutable case_insensitive_map_t<unique_ptr<IcebergPositionalDeleteData>> positional_delete_data;
