+++ src/include/storage/irc_table_set.hpp
@@ -3,7 +3,6 @@
 
 #include "storage/irc_table_entry.hpp"
 #include "storage/iceberg_transaction_data.hpp"
-#include "rest_catalog/objects/add_snapshot_update.hpp"
 
 namespace duckdb {
 struct CreateTableInfo;
@@ -26,8 +25,7 @@ public:
 	IRCAPITableCredentials GetVendedCredentials(ClientContext &context);
 	const string &BaseFilePath() const;
 
-	void Append(IRCTransaction &transaction, vector<IcebergManifestEntry> &&data_files);
-	rest_api_objects::AddSnapshotUpdate CreateSnapshotUpdate(DatabaseInstance &db, ClientContext &context);
+	void AddSnapshot(IRCTransaction &transaction, vector<IcebergManifestEntry> &&data_files);
 
 public:
 	IRCatalog &catalog;
