+++ src/storage/iceberg_insert.cpp
@@ -230,7 +230,7 @@ SinkFinalizeType IcebergInsert::Finalize(Pipeline &pipeline, Event &event, Clien
 	auto &table_info = irc_table.table_info;
 	auto &transaction = IRCTransaction::Get(context, table->catalog);
 
-	table_info.Append(transaction, std::move(global_state.written_files));
+	table_info.AddSnapshot(transaction, std::move(global_state.written_files));
 	transaction.MarkTableAsDirty(irc_table);
 	return SinkFinalizeType::READY;
 }
