+++ src/storage/irc_transaction.cpp
@@ -1,8 +1,10 @@
-#include "storage/irc_transaction.hpp"
-#include "storage/irc_catalog.hpp"
 #include "duckdb/parser/parsed_data/create_view_info.hpp"
 #include "duckdb/catalog/catalog_entry/index_catalog_entry.hpp"
 #include "duckdb/catalog/catalog_entry/view_catalog_entry.hpp"
+
+#include "storage/irc_transaction.hpp"
+#include "storage/irc_catalog.hpp"
+#include "storage/table_update/iceberg_add_snapshot.hpp"
 
 namespace duckdb {
 
@@ -25,10 +27,13 @@ void IRCTransaction::Commit() {
 	context->transaction.BeginTransaction();
 
 	for (auto &table : dirty_tables) {
-		auto update = table->table_info.CreateSnapshotUpdate(db, *context);
-		// - serialize this to JSON
-		// - hit the REST API (POST to '/v1/{prefix}/transactions/commit'), sending along this payload
-		// - profit ???
+		auto &transaction_data = *table->table_info.transaction_data;
+		for (auto &update : transaction_data.updates) {
+			auto table_update = update->CreateUpdate(db, *context);
+			// - serialize this to JSON
+			// - hit the REST API (POST to '/v1/{prefix}/transactions/commit'), sending along this payload
+			// - profit ???
+		}
 	}
 	context->transaction.ClearTransaction();
 }
@@ -37,12 +42,16 @@ void IRCTransaction::CleanupFiles() {
 	// remove any files that were written
 	auto &fs = FileSystem::GetFileSystem(db);
 	for (auto &table : dirty_tables) {
-		auto &table_info = table->table_info;
-		auto &transaction_data = *table_info.transaction_data;
-
-		auto &data_files = transaction_data.manifest_file.data_files;
-		for (auto &data_file : data_files) {
-			fs.TryRemoveFile(data_file.file_path);
+		auto &transaction_data = *table->table_info.transaction_data;
+		for (auto &update : transaction_data.updates) {
+			if (update->type != IcebergTableUpdateType::ADD_SNAPSHOT) {
+				continue;
+			}
+			auto &add_snapshot = update->Cast<IcebergAddSnapshot>();
+			auto &data_files = add_snapshot.manifest_file.data_files;
+			for (auto &data_file : data_files) {
+				fs.TryRemoveFile(data_file.file_path);
+			}
 		}
 	}
 }
